Add blueprint builder

+++ src/ast/type_registry/primitive_type.rs
@@ -36,7 +36,7 @@ impl Display for PrimitiveType {
 }
 
 impl PrimitiveType {
-    pub fn size(&self) -> u32 {
+    pub fn size(&self) -> usize {
         match self {
             PrimitiveType::Bool => 1,
             PrimitiveType::Int => 4,
